loader protocol: hook LoadImage() lazily
Right now we're not hooking the system table's LoadImage(), so that
loaders that depend on /not/ having shim's protocol won't get it.

This means anything they load will fail unless it is signed by something
in the system dbx, and if they want to use shim's keyring they need to
opt in to shim's protocol, and we shouldn't hook it immediately.

Mate mentioned that we probably *do* want to hook it after all to keep
the image handles we've created from getting passed to the system's
LoadImage() later, so in this version we're hooking it on first use.

Signed-off-by: Peter Jones <pjones@redhat.com>