handle_image(): make verification conditional
This changes the verification path in the loader protocol to keep track
of if there is a parent image that has already been verified (currently
always false) and if so, to not do any verification in the
handle_image() code path.

Originally the idea was to log the sub-image digest in the tpm log, but
the EFI APIs do not give us a way to do that without also doing the
extend we're trying to avoid.

Signed-off-by: Peter Jones <pjones@redhat.com>