loader-protocol: try to attach loaded image proto to parent handle
Currently UKIs using systemd-stub pass a parent handle without a loaded image
protocol attached, so the cache matching fails.
Try to attach it before use. In case it's already a loaded image, it will fail
and we just use the passed handle. If it works, then we use the new handle.

This allows using unchanged UKIs with shim 16.

Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>